fix(framework): Hardcode json-schema-faker
Json-schema-faker is causing big HMR and Webpack headaches when @novu/framework is used in Next.js.
To address the issue, we decided to go old-school and hardcode the IIFE version of the source code in our package.
The code was copied for https://unpkg.com/browse/json-schema-faker@0.5.6/dist/main.iife.js.

PLEASE NOTE THE CODE WAS SLIGHTLY MODIFIED TO MAKE IT WORK IN @novu/framework. See the end of this file.

See https://github.com/json-schema-faker/json-schema-faker/issues/796#issuecomment-2433335751.